+++ src/node/config.ts
@@ -20,9 +20,14 @@ import {
   createEsbuildPlugin,
   createEsbuildRenderChunkPlugin
 } from './build/buildPluginEsbuild'
-import { ServerPlugin } from './server'
+import { BuildPlugin } from './build'
+import { Context, ServerPlugin } from './server'
 import { Resolver, supportedExts } from './resolver'
-import { Transform, CustomBlockTransform } from './transform'
+import {
+  Transform,
+  CustomBlockTransform,
+  IndexHtmlTransform
+} from './transform'
 import { DepOptimizationOptions } from './optimizer'
 import { ServerOptions } from 'https'
 import { lookupFile } from './utils'
@@ -38,6 +43,40 @@ export type PreprocessOptions = SFCStyleCompileOptions['preprocessOptions']
 export type CssPreprocessOptions = Partial<
   Record<PreprocessLang, PreprocessOptions>
 >
+
+/**
+ * https://github.com/koajs/cors#corsoptions
+ */
+export interface CorsOptions {
+  /**
+   * `Access-Control-Allow-Origin`, default is request Origin header
+   */
+  origin?: string | ((ctx: Context) => string)
+  /**
+   * `Access-Control-Allow-Methods`, default is 'GET,HEAD,PUT,POST,DELETE,PATCH'
+   */
+  allowMethods?: string | string[]
+  /**
+   * `Access-Control-Expose-Headers`
+   */
+  exposeHeaders?: string | string[]
+  /**
+   * `Access-Control-Allow-Headers`
+   */
+  allowHeaders?: string | string[]
+  /**
+   * `Access-Control-Max-Age` in seconds
+   */
+  maxAge?: string | number
+  /**
+   * `Access-Control-Allow-Credentials`, default is false
+   */
+  credentials?: boolean | ((ctx: Context) => boolean)
+  /**
+   * Add set headers to `err.header` if an error is thrown
+   */
+  keepHeadersOnError?: boolean
+}
 
 export { Resolver, Transform }
 
@@ -80,6 +119,10 @@ export interface SharedConfig {
    * Custom file transforms.
    */
   transforms?: Transform[]
+  /**
+   * Custom index.html transforms.
+   */
+  indexHtmlTransforms?: IndexHtmlTransform[]
   /**
    * Define global variable replacements.
    * Entries will be defined on `window` during dev and replaced during build.
@@ -219,6 +262,13 @@ export interface ServerConfig extends SharedConfig {
    */
   proxy?: Record<string, string | ProxiesOptions>
   /**
+   * Configure CORS for the dev server.
+   * Uses [@koa/cors](https://github.com/koajs/cors).
+   * Set to `true` to allow all methods from any origin, or configure separately
+   * using an object.
+   */
+  cors?: CorsOptions | boolean
+  /**
    * A plugin function that configures the dev server. Receives a server plugin
    * context object just like the internal server plugins. Can also be an array
    * of multiple server plugin functions.
@@ -226,57 +276,63 @@ export interface ServerConfig extends SharedConfig {
   configureServer?: ServerPlugin | ServerPlugin[]
 }
 
-export interface BuildConfig extends SharedConfig {
+export interface BuildConfig extends Required<SharedConfig> {
+  /**
+   * Entry. Use this to specify a js entry file in use cases where an
+   * `index.html` does not exist (e.g. serving vite assets from a different host)
+   * @default 'index.html'
+   */
+  entry: string
   /**
    * Base public path when served in production.
    * @default '/'
    */
-  base?: string
+  base: string
   /**
    * Directory relative from `root` where build output will be placed. If the
    * directory exists, it will be removed before the build.
    * @default 'dist'
    */
-  outDir?: string
+  outDir: string
   /**
    * Directory relative from `outDir` where the built js/css/image assets will
    * be placed.
    * @default '_assets'
    */
-  assetsDir?: string
+  assetsDir: string
   /**
    * Static asset files smaller than this number (in bytes) will be inlined as
    * base64 strings. Default limit is `4096` (4kb). Set to `0` to disable.
    * @default 4096
    */
-  assetsInlineLimit?: number
+  assetsInlineLimit: number
   /**
    * Whether to code-split CSS. When enabled, CSS in async chunks will be
    * inlined as strings in the chunk and inserted via dynamically created
    * style tags when the chunk is loaded.
    * @default true
    */
-  cssCodeSplit?: boolean
+  cssCodeSplit: boolean
   /**
    * Whether to generate sourcemap
    * @default false
    */
-  sourcemap?: boolean
+  sourcemap: boolean | 'inline'
   /**
    * Set to `false` to disable minification, or specify the minifier to use.
    * Available options are 'terser' or 'esbuild'.
    * @default 'terser'
    */
-  minify?: boolean | 'terser' | 'esbuild'
+  minify: boolean | 'terser' | 'esbuild'
   /**
    * The option for `terser`
    */
-  terserOptions?: RollupTerserOptions
+  terserOptions: RollupTerserOptions
   /**
    * Transpile target for esbuild.
    * @default 'es2020'
    */
-  esbuildTarget?: string
+  esbuildTarget: string
   /**
    * Build for server-side rendering, only as a CLI flag
    * for programmatic usage, use `ssrBuild` directly.
@@ -290,54 +346,77 @@ export interface BuildConfig extends SharedConfig {
    *
    * https://rollupjs.org/guide/en/#big-list-of-options
    */
-  rollupInputOptions?: ViteRollupInputOptions
+  rollupInputOptions: ViteRollupInputOptions
   /**
    * Will be passed to bundle.generate()
    *
    * https://rollupjs.org/guide/en/#big-list-of-options
    */
-  rollupOutputOptions?: RollupOutputOptions
+  rollupOutputOptions: RollupOutputOptions
   /**
    * Will be passed to rollup-plugin-vue
    *
    * https://github.com/vuejs/rollup-plugin-vue/blob/next/src/index.ts
    */
-  rollupPluginVueOptions?: Partial<RollupPluginVueOptions>
+  rollupPluginVueOptions: Partial<RollupPluginVueOptions>
   /**
    * Will be passed to @rollup/plugin-node-resolve
    * https://github.com/rollup/plugins/tree/master/packages/node-resolve#dedupe
    */
-  rollupDedupe?: string[]
+  rollupDedupe: string[]
   /**
    * Whether to log asset info to console
    * @default false
    */
-  silent?: boolean
+  silent: boolean
   /**
    * Whether to write bundle to disk
    * @default true
    */
-  write?: boolean
+  write: boolean
   /**
    * Whether to emit index.html
    * @default true
    */
-  emitIndex?: boolean
+  emitIndex: boolean
   /**
    * Whether to emit assets other than JavaScript
    * @default true
    */
-  emitAssets?: boolean
+  emitAssets: boolean
+  /**
+   * Whether to emit a manifest.json under assets dir to map hash-less filenames
+   * to their hashed versions. Useful when you want to generate your own HTML
+   * instead of using the one generated by Vite.
+   *
+   * Example:
+   *
+   * ```json
+   * {
+   *   "main.js": "main.68fe3fad.js",
+   *   "style.css": "style.e6b63442.css"
+   * }
+   * ```
+   * @default false
+   */
+  emitManifest?: boolean
   /**
    * Predicate function that determines whether a link rel=modulepreload shall be
    * added to the index.html for the chunk passed in
    */
-  shouldPreload?: (chunk: OutputChunk) => boolean
+  shouldPreload: ((chunk: OutputChunk) => boolean) | null
   /**
    * Enable 'rollup-plugin-vue'
    * @default true
    */
   enableRollupPluginVue?: boolean
+  /**
+   * Plugin functions that mutate the Vite build config. The `builds` array can
+   * be added to if the plugin wants to add another Rollup build that Vite writes
+   * to disk. Return a function to gain access to each build's output.
+   * @internal
+   */
+  configureBuild?: BuildPlugin | BuildPlugin[]
 }
 
 export interface ViteRollupInputOptions extends RollupInputOptions {
@@ -359,7 +438,7 @@ export interface ViteRollupInputOptions extends RollupInputOptions {
   pluginsOptimizer?: RollupPlugin[]
 }
 
-export interface UserConfig extends BuildConfig, ServerConfig {
+export interface UserConfig extends Partial<BuildConfig>, ServerConfig {
   plugins?: Plugin[]
 }
 
@@ -368,8 +447,10 @@ export interface Plugin
     UserConfig,
     | 'alias'
     | 'transforms'
+    | 'indexHtmlTransforms'
     | 'define'
     | 'resolvers'
+    | 'configureBuild'
     | 'configureServer'
     | 'vueCompilerOptions'
     | 'vueTransformAssetUrls'
@@ -381,6 +462,7 @@ export interface Plugin
   > {}
 
 export type ResolvedConfig = UserConfig & {
+  env: DotenvParseOutput
   /**
    * Path of config file.
    */
@@ -389,13 +471,10 @@ export type ResolvedConfig = UserConfig & {
 
 const debug = require('debug')('vite:config')
 
-export async function resolveConfig(
-  mode: string,
-  configPath?: string
-): Promise<ResolvedConfig | undefined> {
+export async function resolveConfig(mode: string, configPath?: string) {
   const start = Date.now()
   const cwd = process.cwd()
-  let config: ResolvedConfig | ((mode: string) => ResolvedConfig) | undefined
+
   let resolvedPath: string | undefined
   let isTS = false
   if (configPath) {
@@ -421,21 +500,20 @@ export async function resolveConfig(
   }
 
   try {
+    let userConfig: UserConfig | ((mode: string) => UserConfig) | undefined
+
     if (!isTS) {
       try {
-        config = require(resolvedPath)
+        userConfig = require(resolvedPath)
       } catch (e) {
-        if (
-          !/Cannot use import statement|Unexpected token 'export'|Must use import to load ES Module/.test(
-            e.message
-          )
-        ) {
+        const ignored = /Cannot use import statement|Unexpected token 'export'|Must use import to load ES Module/
+        if (!ignored.test(e.message)) {
           throw e
         }
       }
     }
 
-    if (!config) {
+    if (!userConfig) {
       // 2. if we reach here, the file is ts or using es import syntax, or
       // the user has type: "module" in their package.json (#917)
       // transpile es import syntax to require syntax using rollup.
@@ -465,11 +543,18 @@ export async function resolveConfig(
         format: 'cjs'
       })
 
-      config = await loadConfigFromBundledFile(resolvedPath, code)
+      userConfig = await loadConfigFromBundledFile(resolvedPath, code)
     }
 
-    if (typeof config === 'function') {
-      config = config(mode)
+    let config = (typeof userConfig === 'function'
+      ? userConfig(mode)
+      : userConfig) as ResolvedConfig
+
+    // resolve plugins
+    if (config.plugins) {
+      for (const plugin of config.plugins) {
+        config = resolvePlugin(config, plugin) as any
+      }
     }
 
     // normalize config root to absolute
@@ -483,16 +568,10 @@ export async function resolveConfig(
       )
     }
 
-    // resolve plugins
-    if (config.plugins) {
-      for (const plugin of config.plugins) {
-        config = resolvePlugin(config, plugin)
-      }
-    }
-
+    const env = loadEnv(mode, config.root || cwd)
     config.env = {
       ...config.env,
-      ...loadEnv(mode, config.root || cwd)
+      ...env
     }
     debug(`config resolved in ${Date.now() - start}ms`)
 
@@ -544,10 +623,18 @@ function resolvePlugin(config: UserConfig, plugin: Plugin): UserConfig {
       ...config.define
     },
     transforms: [...(config.transforms || []), ...(plugin.transforms || [])],
+    indexHtmlTransforms: [
+      ...(config.indexHtmlTransforms || []),
+      ...(plugin.indexHtmlTransforms || [])
+    ],
     resolvers: [...(config.resolvers || []), ...(plugin.resolvers || [])],
     configureServer: ([] as ServerPlugin[]).concat(
       config.configureServer || [],
       plugin.configureServer || []
+    ),
+    configureBuild: ([] as BuildPlugin[]).concat(
+      config.configureBuild || [],
+      plugin.configureBuild || []
     ),
     vueCompilerOptions: {
       ...config.vueCompilerOptions,
@@ -623,7 +710,7 @@ function mergeObjectOptions(to: any, from: any) {
   return res
 }
 
-function loadEnv(mode: string, root: string): Record<string, string> {
+export function loadEnv(mode: string, root: string, prefix = 'VITE_') {
   if (mode === 'local') {
     throw new Error(
       `"local" cannot be used as a mode name because it conflicts with ` +
@@ -633,8 +720,7 @@ function loadEnv(mode: string, root: string): Record<string, string> {
 
   debug(`env mode: ${mode}`)
 
-  const nodeEnv = process.env
-  const clientEnv: Record<string, string> = {}
+  const env: DotenvParseOutput = {}
   const envFiles = [
     /** mode local file */ `.env.${mode}.local`,
     /** mode file */ `.env.${mode}`,
@@ -645,32 +731,28 @@ function loadEnv(mode: string, root: string): Record<string, string> {
   for (const file of envFiles) {
     const path = lookupFile(root, [file], true)
     if (path) {
-      const result = dotenv.config({
-        debug: !!process.env.DEBUG || undefined,
-        path
+      const parsed = dotenv.parse(fs.readFileSync(path), {
+        debug: !!process.env.DEBUG || undefined
       })
-      if (result.error) {
-        throw result.error
-      }
-      dotenvExpand(result)
-      for (const key in result.parsed) {
-        const value = (nodeEnv[key] = result.parsed![key])
-        // only keys that start with VITE_ are exposed.
-        if (key.startsWith(`VITE_`)) {
-          clientEnv[key] = value
-        }
-        // set NODE_ENV under a different key so that we know this is set from
-        // vite-loaded .env files. Some users may have default NODE_ENV set in
-        // their system.
-        if (key === 'NODE_ENV') {
-          nodeEnv.VITE_ENV = value
+
+      // let environment variables use each other
+      dotenvExpand({
+        parsed,
+        // prevent process.env mutation
+        ignoreProcessEnv: true
+      } as any)
+
+      // only keys that start with prefix are exposed.
+      for (const [key, value] of Object.entries(parsed)) {
+        if (key.startsWith(prefix) && env[key] === undefined) {
+          env[key] = value
         }
       }
     }
   }
 
-  debug(`env: %O`, clientEnv)
-  return clientEnv
+  debug(`env: %O`, env)
+  return env
 }
 
 // TODO move this into Vue plugin when we extract it