+++ src/node/server/serverPluginModuleResolve.ts
@@ -4,7 +4,11 @@ import fs from 'fs-extra'
 import { ServerPlugin } from '.'
 import { resolveVue } from '../utils'
 import { URL } from 'url'
-import { resolveOptimizedModule, resolveNodeModuleFile } from '../resolver'
+import {
+  resolveOptimizedModule,
+  resolveNodeModuleFile,
+  resolveNodeModule
+} from '../resolver'
 
 const debug = require('debug')('vite:resolve')
 
@@ -69,23 +73,15 @@ export const moduleResolvePlugin: ServerPlugin = ({ root, app, resolver }) => {
     }
 
     const importerFilePath = importer ? resolver.requestToFile(importer) : root
-    const nodeModulePath = resolveNodeModuleFile(importerFilePath, id)
-    if (nodeModulePath) {
-      return serve(id, nodeModulePath, 'node_modules')
+    // #829 node package has sub-package(has package.json), should check it before `resolveNodeModuleFile`
+    const nodeModuleInfo = resolveNodeModule(root, id, resolver)
+    if (nodeModuleInfo) {
+      return serve(id, nodeModuleInfo.entryFilePath!, 'node_modules')
     }
 
-    // resolve relative path request inside unoptimized package while the importer is lack of /index.js
-    // eg.
-    // id: /rewrite-unoptimized-test-package/es/foo
-    // importer: /rewrite-unoptimized-test-package/es/nested
-    // The `id` is not resolved, because it should be /rewrite-unoptimized-test-package/es/nested/foo
-    if (importer) {
-      const relativePath = path.relative(importer.replace(moduleRE, ''), id)
-      const filePath = path.join(importerFilePath, relativePath)
-      const fileRealPath = resolveNodeModuleFile(root, filePath)
-      if (fileRealPath) {
-        return serve(id, fileRealPath, 'node_modules')
-      }
+    const nodeModuleFilePath = resolveNodeModuleFile(importerFilePath, id)
+    if (nodeModuleFilePath) {
+      return serve(id, nodeModuleFilePath, 'node_modules')
     }
 
     if (isMapFile && importer) {