+++ test/test.js
@@ -3,7 +3,6 @@ const path = require('path')
 const execa = require('execa')
 const puppeteer = require('puppeteer')
 const moment = require('moment')
-const { EPERM } = require('constants')
 
 jest.setTimeout(100000)
 
@@ -735,7 +734,32 @@ describe('vite', () => {
     test('optional chaining syntax support', async () => {
       expect(await getText('.syntax')).toBe('baz')
     })
+
+    test('optimizing commonjs dependencies and do named-import from it', async () => {
+      expect((await getText('.cjs-dep-named-export-static')).trim()).toBe(
+        'static import result: success'
+      )
+      await click('.cjs-dep-named-export-dynamic-load')
+      await expectByPolling(
+        async () =>
+          ((await getText('.cjs-dep-named-export-dynamic')) || '').trim(),
+        'dynamic import result: success'
+      )
+    })
   }
+
+  describe('build (multi)', () => {
+    beforeAll(async () => {
+      const buildOutput = await execa(binPath, ['build'], {
+        cwd: path.join(tempDir, 'multi-build')
+      })
+      expect(buildOutput.stdout).toMatch('Build completed')
+      expect(buildOutput.stderr).toBe('')
+    })
+
+    test.todo('index.html renders "hello world"')
+    test.todo('index.mobile.html renders "hello mobile"')
+  })
 
   describe('build', () => {
     let staticServer
@@ -791,6 +815,16 @@ describe('vite', () => {
       // should be inside the async chunk
       expect(code).toMatch(colorToMatch)
     })
+
+    test('build manifest', async () => {
+      const manifest = JSON.parse(
+        await fs.readFile(path.join(tempDir, 'dist/_assets/manifest.json'))
+      )
+      const indexPath = manifest['index.js']
+      expect(
+        await fs.stat(path.join(tempDir, `dist/_assets/${indexPath}`))
+      ).toBeTruthy()
+    })
   })
 
   describe('dev', () => {
@@ -825,7 +859,7 @@ describe('vite', () => {
     declareTests(false)
 
     test('hmr (index.html full-reload)', async () => {
-      expect(await getText('title')).toMatch('Vite App')
+      expect(await getText('title')).toMatch('Vite Playground')
       // hmr
       const reload = page.waitForNavigation({
         waitUntil: 'domcontentloaded'
@@ -834,12 +868,12 @@ describe('vite', () => {
         content.replace('Vite App', 'Vite App Test')
       )
       await reload
-      await expectByPolling(() => getText('title'), 'Vite App Test')
+      await expectByPolling(() => getText('title'), 'Vite Playground Test')
     })
 
     test('hmr (html full-reload)', async () => {
       await page.goto('http://localhost:3000/test.html')
-      expect(await getText('title')).toMatch('Vite App')
+      expect(await getText('title')).toMatch('Vite Playground')
       // hmr
       const reload = page.waitForNavigation({
         waitUntil: 'domcontentloaded'
@@ -848,7 +882,7 @@ describe('vite', () => {
         content.replace('Vite App', 'Vite App Test')
       )
       await reload
-      await expectByPolling(() => getText('title'), 'Vite App Test')
+      await expectByPolling(() => getText('title'), 'Vite Playground Test')
     })
 
     // Assert that all edited files are reflected on page reload